Move detection models to domain module
Relocated the Detection-related classes from `track.model` to `domain` module for better logical separation and organization. Updated all imports and references across the codebase to reflect this change.
