+++ OTVision/track/model/filebased/frame_chunk.py
@@ -6,6 +6,7 @@ from typing import Sequence
 from tqdm import tqdm
 
 from OTVision.dataformat import FRAME, INPUT_FILE_PATH, TRACK_ID
+from OTVision.domain.detection import TrackId
 from OTVision.track.model.filebased.frame_group import FrameGroup, get_output_file
 from OTVision.track.model.frame import (
     FinishedFrame,
@@ -13,7 +14,6 @@ from OTVision.track.model.frame import (
     FrameNo,
     IsLastFrame,
     TrackedFrame,
-    TrackId,
 )
