+++ OTVision/track/tracker/filebased_tracking.py
@@ -6,6 +6,7 @@ from more_itertools import peekable
 from tqdm import tqdm
 
 from OTVision.config import CONFIG, DEFAULT_FILETYPE, OVERWRITE, TRACK
+from OTVision.domain.detection import TrackId
 from OTVision.helpers.log import LOGGER_NAME
 from OTVision.track.model.filebased.frame_chunk import (
     ChunkParser,
@@ -14,13 +15,7 @@ from OTVision.track.model.filebased.frame_chunk import (
     TrackedChunk,
 )
 from OTVision.track.model.filebased.frame_group import FrameGroup, FrameGroupParser
-from OTVision.track.model.frame import (
-    Frame,
-    FrameNo,
-    IsLastFrame,
-    TrackedFrame,
-    TrackId,
-)
+from OTVision.track.model.frame import Frame, FrameNo, IsLastFrame, TrackedFrame
 from OTVision.track.model.tracking_interfaces import (
     ID_GENERATOR,
     Tracker,
